+++ Assets/Scripts/Interfaces/ICanOpenDoor.cs
@@ -3,6 +3,6 @@ namespace Veganimus.Platformer
 {
     public interface ICanOpenDoor
     {
-        byte MaxDoorLevel { get; set; }
+        int MaxDoorLevel { get; set; }
     }
 }
